Tweak some settings and fix a bug I created changing data types

+++ OpenSourceLights/aa_light_setup.h
@@ -69,17 +69,17 @@
 #include <Arduino.h>
 #include <avr/pgmspace.h>
 
-extern const byte NA;
-extern const byte BLINK;
-extern const byte FASTBLINK;
-extern const byte SOFTBLINK;
-extern const byte DIM;
-extern const byte FADEOFF;
-extern const byte XENON;
-extern const byte BACKFIRE;
+extern const uint16_t NA;
+extern const uint16_t BLINK;
+extern const uint16_t FASTBLINK;
+extern const uint16_t SOFTBLINK;
+extern const uint16_t DIM;
+extern const uint16_t FADEOFF;
+extern const uint16_t XENON;
+extern const uint16_t BACKFIRE;
 extern const byte ON;
 extern const byte OFF;
 extern const byte YES;
 extern const byte NO;
 
-extern const byte Schemes[NumSchemes][NumLights][NumStates] PROGMEM;
+extern const uint16_t Schemes[NumSchemes][NumLights][NumStates] PROGMEM;
